pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "flake8>=5.0.4",
    "flake8-blind-except>=0.2.1",
    "flake8-builtins>=2.1.0",
    "flake8-broken-line>=0.6.0",
    "flake8-bugbear>=23.1.20",
    "flake8-comprehensions>=3.10.1",
    "flake8-eradicate>=1.4.0",
    "flake8-executable>=2.1.3",
    "flake8-fixme>=1.1.1",
    "flake8-mutable>=1.2.0",
    "flake8-pep3101>=2.0.0",
    "flake8-pie>=0.16.0",
    "flake8-print>=5.0.0",
    "flake8-tidy-imports>=4.8.0",
    "flake8-bandit>=4.1.1",
    "pep8-naming>=0.13.3",
    "mypy>=0.991",
    "pytest>=7.2.1",
    "pytest-aiohttp>=1.0.4",
    "asynctest>=0.13.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "frolin28", email = "frolkin2801@gmail.com"},
]
dependencies = [
    "aiohttp>=3.8.3",
    "cchardet>=2.1.7",
    "aiodns>=3.0.0",
    "aiohttp-autoreload>=0.0.1",
    "uvloop>=0.17.0",
    "trafaret>=2.1.1",
    "ruamel.yaml>=0.17.21",
    "aiohttp-security>=0.4.0",
    "umongo>=3.1.0",
    "motor>=3.1.1",
    "packaging>=23.0",
    "aiohttp-session[secure]>=2.12.0",
    "passlib>=1.7.4",
]
requires-python = ">=3.10"
license = {text = "MIT"}
